airflow/requirements.txt
certifi
charset-normalizer
docopt
hdfs
idna
numpy
pandas
pyarrow
python-dateutil
pytz
requests
six
sodapy
tzdata
urllib3
hdfs
sqlalchemy
python-dotenv
duckdb
pyspark
plotly
scikit-learn
xgboost
